Consistency, all layers now named PixiLayerSomething
