+++ modules/pixi/PixiLayerImproveOsm.js
@@ -6,16 +6,25 @@ import { PixiLayer } from './PixiLayer';
 import { getIconSprite } from './helpers';
 
 
-const LAYERID = 'osmose';
+const LAYERID = 'improveOSM';
 const LAYERZINDEX = 10;
 const MINZOOM = 12;
 
+// A mapping of improveOSM rule numbers and their respective tint colors.
+const TINTS = new Map();
+TINTS.set('tr', 0xec1c24);         // turn restrictions
+TINTS.set('ow', 0x1e90ff);         // oneway restrictions
+TINTS.set('mr-road', 0xb452cd);    // missing missing road
+TINTS.set('mr-path', 0xa0522d);    // missing path
+TINTS.set('mr-parking', 0xeeee00); // missing parking
+TINTS.set('mr-both', 0xffa500);    // missing road + parking
+
 
 /**
- * PixiOsmose
+ * PixiLayerImproveOsm
  * @class
  */
-export class PixiOsmose extends PixiLayer {
+export class PixiLayerImproveOsm extends PixiLayer {
 
   /**
    * @constructor
@@ -32,10 +41,12 @@ export class PixiOsmose extends PixiLayer {
     this._service = null;
     this.getService();
 
+    // var throttledRedraw = _throttle(function () { this.dispatch.call('change'); }, 1000);
+
     // Create marker texture
     this.textures = {};
     const marker = new PIXI.Graphics()
-      .lineStyle(1, 0x33333)
+      .lineStyle(1, 0x333333)
       .beginFill(0xffffff)
       .drawPolygon([16,3, 4,3, 1,6, 1,17, 4,20, 7,20, 10,27, 13,20, 16,20, 19,17.033, 19,6])
       .endFill()
@@ -43,7 +54,7 @@ export class PixiOsmose extends PixiLayer {
 
     const renderer = context.pixi.renderer;
     const options = { resolution: 2 };
-    this.textures.osmoseMarker = renderer.generateTexture(marker, options);
+    this.textures.improveOSMMarker = renderer.generateTexture(marker, options);
   }
 
 
@@ -52,10 +63,10 @@ export class PixiOsmose extends PixiLayer {
    * to gain access to them, and bind any event handlers a single time.
    */
   getService() {
-    if (services.osmose && !this._service) {
-      this._service = services.osmose;
-      // this._service.event.on('loadedImages', throttledRedraw);
-    } else if (!services.osmose && this._service) {
+    if (services.improveOSM && !this._service) {
+      this._service = services.improveOSM;
+      // this._service.event.on('loaded', throttledRedraw);
+    } else if (!services.improveOSM && this._service) {
       this._service = null;
     }
 
@@ -75,20 +86,19 @@ export class PixiOsmose extends PixiLayer {
     const service = this.getService();
     if (!service) return;
 
-    const visibleData = service.getItems(context.projection);
+    const visibleData = service.getItems(context.projection);  // note: context.projection !== pixi projection
     visibleData.forEach(d => {
       const featureID = `${LAYERID}-${d.id}`;
       let feature = featureCache.get(featureID);
 
       if (!feature) {
-        const marker = new PIXI.Sprite(this.textures.osmoseMarker);
+        const marker = new PIXI.Sprite(this.textures.improveOSMMarker);
         marker.name = featureID;
         marker.buttonMode = true;
         marker.interactive = true;
         marker.zIndex = -d.loc[1];   // sort by latitude ascending
         marker.anchor.set(0.5, 1);   // middle, bottom
-        const color = service.getColor(d.item);
-        marker.tint = PIXI.utils.string2hex(color);
+        marker.tint = TINTS.get(d.itemType) || 0xffffff;
         this.container.addChild(marker);
 
         if (d.icon) {
